Call supervisors before other listeners
Calling supervisors first makes ``emitter.on(console.log)`` print the events in
the order they are emitted.

The old behavior of calling listeners first was not specified by any docs or
tests. If supervisors are called after the other listeners, then events emitted
by a listener will be printed before the event that triggered the listener in
the first place. Event debugging using supervisors with this behavior can be
rather confusing.